keypoint matching by sub-regions
The center of the image is divided into a certain number of 'stripes'. Keypoints are searched inside the  center of each frame and then assigned to the region they belong to. Then, keypoints matching is performed per stripe between each couple of corresponding frames.
TODO: assign keypoints to a structure in which each element contains keypoints of a certain region.